mcp_system/requirements.txt
google-genai>=0.8.0
fastmcp>=3.2.0
mcp>=1.27.0
python-dotenv>=1.0.0
